% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_categories.R
\name{decode_categories}
\alias{decode_categories}
\title{Decode Categories}
\usage{
decode_categories(encoded_values, prime_keys)
}
\arguments{
\item{encoded_values}{Values encoded by \code{encode_categories()}.}

\item{prime_keys}{Keys generated by \code{make_prime_keys()}.}
}
\value{
A list of lists with decoded categories.
}
\description{
Decodes categories encoded by \code{encode_categories()}, using a key generated by \code{make_prime_keys()}.
}
\examples{
list_of_lists <- decode_categories(encoded_values, prime_keys)
}
\seealso{
\code{\link[=encode_categories]{encode_categories()}}, \code{\link[=make_prime_kesy]{make_prime_kesy()}}
}
