% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_categories.R
\name{encode_categories}
\alias{encode_categories}
\title{Encode Categories}
\usage{
encode_categories(categorical_variable, prime_keys)
}
\arguments{
\item{categorical_variable}{The values to be encoded.}

\item{prime_keys}{An encoding key, as returned by \code{make_prime_keys()}.}
}
\value{
A list of encoded values.
}
\description{
Encodes a list of categorical values for storage in a single field using prime number multiplication.
}
\examples{
encoded_values <- encode_categories(shifts, prime_keys)
}
\seealso{
\code{\link[=make_prime_keys]{make_prime_keys()}}
}
