% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_dates.R
\name{extract_dates}
\alias{extract_dates}
\title{Extract Dates}
\usage{
extract_dates(schedule_table, first_year)
}
\arguments{
\item{schedule_table}{A schedule table extracted bu \code{extract_table()}}

\item{first_year}{The first year of the dates, as found by \code{find_first_year()}.}
}
\value{
A tibble of the dates in the schedule.
}
\description{
Extracts the dates from a schedule extracted by \code{extract_table()}, setting the year of all the dates to the same year.
}
\examples{
dates <- extract_dates(schedule_data, 2022)
}
\seealso{
\code{\link[=extract_table]{extract_table()}}, \code{\link[=find_first_year]{find_first_year()}}
}
