% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_doctor_names.R
\name{extract_doctor_names}
\alias{extract_doctor_names}
\title{Extract Doctor Names}
\usage{
extract_doctor_names(table)
}
\arguments{
\item{table}{A table as returned by `extract_table()` and indexed by
`find_largest_table()`.}
}
\value{
A tibble of doctors’ names.
}
\description{
Returns a tibble of the doctors’ names from the schedule.
}
\examples{
doctor_names <- extract_doctor_names(table)
}
\seealso{
[medinetparser::extract_table()],
  [medinetparser::find_largest_table()]
}
