% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tables.R
\name{extract_tables}
\alias{extract_tables}
\title{Extract Tables}
\usage{
extract_tables(file_path)
}
\arguments{
\item{file_path}{The hath to the HTML file downloaded from Medinet.}
}
\value{
A list of HTML tables for use with rvest.
}
\description{
Loads the HTML file downloaded from Medinet and returns the HTML tables for
further analysis. NB! This takes a while to run. Be patient.
}
\examples{
tables <- extract_tables('data.html')
}
