% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tidy_schedule.R
\name{extract_tidy_schedule}
\alias{extract_tidy_schedule}
\title{Extract Tidy Schedule}
\usage{
extract_tidy_schedule(file_path)
}
\arguments{
\item{file_path}{File path to HTML file from Medinet.}
}
\value{
A tidy tibble with schedule data.
}
\description{
Takes the name of an HTML file from Medinet and outputs a tidy tibble with the schedule data.
}
\examples{
tidy_schedule <- extract_tidy_schedule('./schedule.html')
}
