% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_date_limits.R
\name{find_date_limits}
\alias{find_date_limits}
\title{Find Date Limits}
\usage{
find_date_limits(table, first_year_of_schedule)
}
\arguments{
\item{table}{Schedule table as found by \code{extract_tables()}.}

\item{first_year_of_schedule}{First year of schedule as found by
\code{find_first_year()}.}
}
\value{
Outputs a list consisting of two lubridate dates.
}
\description{
Finds the first and last date of the schedule.
}
\examples{
dates <- find_date_limits(table, '2022')
}
\seealso{
\code{\link[=extract_tables]{extract_tables()}},
\code{\link[=find_first_year]{find_first_year()}}
}
