% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_first_year.R
\name{find_first_year}
\alias{find_first_year}
\title{Find First Year}
\usage{
find_first_year(tables, first_week_number)
}
\arguments{
\item{tables}{Tables output by \code{extract_tables()}.}

\item{first_week_number}{Week number in format "v \link{week number} as output by \code{find_week_number()}.}
}
\value{
The first year of the schedule as a character variable.
}
\description{
Find the first year in the schedule.
}
\examples{
first_year <- find_first_year(tables, 'v 7')
}
\seealso{
\code{\link[=extract_tables]{extract_tables()}}, \code{\link[=find_week_number]{find_week_number()}}
}
