% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_largest_table.R
\name{find_largest_table}
\alias{find_largest_table}
\title{Find Largest Table}
\usage{
find_largest_table(tables)
}
\arguments{
\item{tables}{HTML tables imported by \emph{extract_table}.}
}
\value{
An integer indicating which table has the largest area.
}
\description{
Finds the largest table among the ones input from Medinet, as this is the
most likely to contain the schedule data.
}
\examples{
largest_table_index <- find_largest_table(tables)
}
\seealso{
\code{\link[=extract_table]{extract_table()}}
}
