% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_week_number.R
\name{find_week_number}
\alias{find_week_number}
\title{Find Week Number}
\usage{
find_week_number(table)
}
\arguments{
\item{table}{Schedule found using \code{extract_tables()} and \code{find_largest_table()}.}
}
\value{
First week number of schedule in format "v \link{week number}".
}
\description{
Finds the first week number in the schedule and returns "v \link{week number}".
}
\examples{
week_number <- find_week_number(tables[[19]])
}
\seealso{
\code{\link[=extract_tables]{extract_tables()}}, \code{\link[=find_largest_table]{find_largest_table()}}
}
