% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_primes.R
\name{generate_primes}
\alias{generate_primes}
\title{Generate Primes}
\usage{
generate_primes(number_of_primes)
}
\arguments{
\item{number_of_primes}{The number of prime numbers needed.}
}
\value{
A list of prime numbers.
}
\description{
Generates a list of prime numbers starting at 2, to be used for encoding.
}
\examples{
list_of_prime_numbers <- generate_primes(15)
}
