% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_min_max_dates.R
\name{get_min_max_dates}
\alias{get_min_max_dates}
\title{Get Min Max Dates}
\usage{
get_min_max_dates(tidy_schedule)
}
\arguments{
\item{tidy_schedule}{A schedule as returned by \code{load_tidy_schedule()}}
}
\value{
A list where the first value is the first date in the schedule and
the last value is the last date in the schedule.
}
\description{
Returns a list of the first and last date in a schedule as returned by
\code{load_tidy_schedule()}. Meant to be used for creating titles for plots.
}
\examples{
min_max_dates <- get_min_max_dates(tidy_schedule)
}
\seealso{
\code{\link[=load_tidy_schedule]{load_tidy_schedule()}}
}
