% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_tidy_schedule.R
\name{load_tidy_schedule}
\alias{load_tidy_schedule}
\title{Load Tidy Schedule}
\usage{
load_tidy_schedule(file_path)
}
\arguments{
\item{file_path}{String containing the path to the HTML file.}
}
\value{
A tidy tibble containing the schedule data for analysis.
}
\description{
Loads a HTML file from Medinet and returns the schedule as a tidy tibble.
Note that the function might take a few minutes to run on large schedule
files.
}
\examples{
schedule_tibble <- load_tidy_schedule('file_from_medinet.html')
}
