% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_schedule_files.R
\name{merge_schedule_files}
\alias{merge_schedule_files}
\title{Merge Schedule Files}
\usage{
merge_schedule_files(list_of_file_paths)
}
\arguments{
\item{list_of_file_paths}{A list of paths to HTML files downloaded from
Medinet.}
}
\value{
A tibble like that returned by \code{load_tidy_schedule()}.
}
\description{
Takes a list of paths to HTML files downloaded from Medinet, imports them
using \code{load_tidy_schedule()} and merges them for analysis, removing duplicate
entries.
}
\examples{
\dontrun{
c('path/to/first/file.html', 'path/to/second/file.html') |>
  merge_schedule_files()}
}
\seealso{
\code{\link[=load_tidy_schedule]{load_tidy_schedule()}}
}
