% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shift_type_by_frequency.R
\name{plot_shift_type_by_frequency}
\alias{plot_shift_type_by_frequency}
\title{Plot Shift Type by Frequency}
\usage{
plot_shift_type_by_frequency(tidy_schedule)
}
\arguments{
\item{tidy_schedule}{A tibble containing a schedule, as loaded by
\code{load_tidy_schedule()}.}
}
\value{
A ggplot with a bar chart.
}
\description{
Plots a bar graph of shift type frequencies.
}
\examples{
# Display graph right away.
plot_shift_type_by_frequency(tidy_schedule)

# Plot shifts for a single doctor, collapsing related categories,
# saving the ggplot to an object for later use.
saved_graph <- tidy_schedule \%>\%
    dplyr::filter(doctor_name == 'Åkesson, Annika') \%>\%
    dplyr::mutate(shift_type = forcats::fct_collapse(shift_type, `C-op` = c('Opkir', 'Opkk'))) \%>\%
    plot_shift_type_by_frequency()
}
\seealso{
\code{\link[=load_tidy_schedule]{load_tidy_schedule()}}
}
