% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shifts_per_doctor.R
\name{plot_shifts_per_doctor}
\alias{plot_shifts_per_doctor}
\title{Plot Shifts Per Doctor}
\usage{
plot_shifts_per_doctor(tidy_schedule)
}
\arguments{
\item{tidy_schedule}{A tibble containing a schedule, as loaded by
\code{load_tidy_schedule()}.}
}
\value{
A ggplot with a stacked bar chart.
}
\description{
Plots a bar graph of the total shifts per doctor, stacked by shift type.
}
\examples{
# Display graph right away.
plot_shifts_per_doctor(example_schedule)

# Plot only a subset of shifts.
example_schedule |>
    dplyr::filter(shift_type \%in\% c('Pjour', 'Bjour')) |>
    plot_shifts_per_doctor()

# Same as above but also change axis label and save plot as
# an object for later use.
plot_of_shifts <- example_schedule |>
    dplyr::filter(shift_type \%in\% c('Pjour', 'Bjour')) |>
    plot_shifts_per_doctor() +
    ggplot2::ylab('Number of Night Shifts')
}
\seealso{
\code{\link[=load_tidy_schedule]{load_tidy_schedule()}}
}
