% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_shifts.R
\name{sample_shifts}
\alias{sample_shifts}
\title{Sample Shifts}
\usage{
sample_shifts(schedule_tibble, samples = 10)
}
\arguments{
\item{schedule_tibble}{A schedule tibble generated by
\code{load_tidy_schedule()}.}

\item{samples}{The number of samples wanted. Defaults to ten.}
}
\value{
A tibble of samples.
}
\description{
Draws a number of samples from a tibble generated by the medinetparser
package, to check against the original Medinet schedule. This function is
meant to facilitate verifying data accuracy before further analysis. The
reason for this being that Medinet can be configured in different ways
in different departments. During development medinetparser has been
tested against schedules in two departments in separate hospitals, but
compatibility with every possible configuration cannot be guaranteed,
so checking tibbles generated by medinetparser against your own
department’s schedule is recommended.
}
\examples{
sample_of_shifts <- sample_shifts(schedule_tibble)
}
\seealso{
\code{\link[=load_tidy_schedule]{load_tidy_schedule()}}
}
