% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_schedule_row.R
\name{tidy_schedule_row}
\alias{tidy_schedule_row}
\title{Tidy Schedule Row}
\usage{
tidy_schedule_row(schedule_row, encoding_keys, schedule_dates)
}
\arguments{
\item{schedule_row}{A row from \code{extract_table()}.}

\item{encoding_keys}{Encoding keys from \code{make_prime_keys()}.}

\item{schedule_dates}{Dates from \code{extract_dates()}.}
}
\value{
A three column tibble.
}
\description{
Takes a row from the schedule, as extracted from \code{extract_table()}, prime number encoding keys, as generated by \code{make_prime_keys()}, and a list of dates, as generated by \code{extract_dates()}. Returns a tibble with three columns; date, doctor, and shift_type.
}
\examples{
tidy_row <- tidy_schedule_row(row, keys, dates)
}
\seealso{
\code{\link[=extract_table]{extract_table()}}, \code{\link[=make_prime_keys]{make_prime_keys()}}, \code{\link[=extract_dates]{extract_dates()}}
}
