% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tables.R
\name{extract_tables}
\alias{extract_tables}
\title{Extract Tables}
\usage{
extract_tables(file_path)
}
\arguments{
\item{file_path}{The hath to the HTML file downloaded from Medinet.}
}
\value{
A list of HTML tables for use with rvest.
}
\description{
Loads the HTML file downloaded from Medinet and returns the HTML tables for
further analysis. NB! This takes a while to run. Be patient.
}
\details{
If you run into memory problems using this function, exporting fewer weeks
at a time from Medinet is suggested. Tallies can easily be added if you are
interested in analyzing a longer period of time. Tables showing distribution
of shifts over time can be bound using dplyr’s \code{bind_cols()}.
}
\examples{
tables <- extract_tables('data.html')
}
