% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_first_year.R
\name{find_first_year}
\alias{find_first_year}
\title{Find First Year}
\usage{
find_first_year(file_path)
}
\arguments{
\item{file_path}{The file path to the schedule HTML file.}
}
\value{
The first year of the schedule as an integer variable.
}
\description{
Find the first year in the schedule.
}
\examples{
first_year <- find_first_year('./data.html')
}
