% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_prime_keys.R
\name{make_prime_keys}
\alias{make_prime_keys}
\title{Make Prime Keys}
\usage{
make_prime_keys(categories)
}
\arguments{
\item{categories}{A list of categories for which to generate keys.}
}
\value{
A list of prime numbers, named after the input categories.
}
\description{
Generates keys that can be used for encoding several different categorical
values in the same field of a tibble.
}
\examples{
encoding_key <- make_prime_keys(list_of_categories)
}
