% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_difference_from_mean.R
\name{plot_difference_from_mean}
\alias{plot_difference_from_mean}
\title{Plot Difference from Mean}
\usage{
plot_difference_from_mean(tidy_schedule)
}
\arguments{
\item{tidy_schedule}{A tibble containing a schedule, as loaded by
\code{load_tidy_schedule()}.}
}
\value{
A ggplot with a bar chart.
}
\description{
Plots a bar graph of the difference between the number of shifts that a
doctor is scheduled and the mean for all the doctors in the schedule.
}
\examples{
# Show plot right away.
plot_difference_from_mean(tidy_schedule)

# Filter certain shift types, exclude one doctor, and save the ggplot as an object for later use.
saved_graph <- tidy_schedule \%>\%
    filter(shift_type \%in\% c('Pjour', 'Mjour')) \%>\%
    filter(doctor_name != 'Hansson, Anders') \%>\%
    plot_difference_from_mean()
}
\seealso{
\code{\link[=load_tidy_schedule]{load_tidy_schedule()}}
}
