% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_shifts.R
\name{sample_shifts}
\alias{sample_shifts}
\title{Sample Shifts}
\usage{
sample_shifts(schedule_tibble, samples = 10)
}
\arguments{
\item{schedule_tibble}{A schedule tibble generated by
\code{extract_tidy_schedule()}.}

\item{samples}{The number of samples wanted. Defaults to ten.}
}
\value{
A tibble of samples.
}
\description{
Draw a number of samples from a tibble generated by the medinetparser
package, to check against the original Medinet schedule. This function is
meant to facilitate verifying data accuracy before further analysis.
}
\examples{
sample_of_shifts <- sample_shifts(schedule_tibble)
}
\seealso{
\code{\link[=extract_tidy_schedule]{extract_tidy_schedule()}}
}
