% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tally_shifts.R
\name{tally_shifts}
\alias{tally_shifts}
\title{Tally Shifts}
\usage{
tally_shifts(table)
}
\arguments{
\item{table}{A table as extracted by \code{extract_table()} and indexed by \code{find_largest_table()}.}
}
\value{
A tibble of shift types against doctors.
}
\description{
Outputs a tally of different shift types against each doctor in the schedule.
}
\examples{
shift_tallies <- tally_shifts(table[[19]])
}
\seealso{
\code{\link[=extract_table]{extract_table()}}, \code{\link[=find_largest_table]{find_largest_table()}}
}
