% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tally_shifts.R
\name{tally_shifts}
\alias{tally_shifts}
\title{Tally Shifts}
\usage{
tally_shifts(tidy_schedule)
}
\arguments{
\item{tidy_schedule}{A tibble generated by \code{load_tidy_schedule()}.}
}
\value{
A tibble with tallies of shift types per doctor.
}
\description{
Takes a tidy schedule generated by \code{load_tidy_schedule()} and generates a table of tallies per doctor.
}
\examples{
tally_table <- tally_shifts(tidy_schedule)
}
\seealso{
\code{\link[=load_tidy_schedule]{load_tidy_schedule()}}
}
